Changes method of rounding.

+++ core/src/main/java/com/graphhopper/storage/CHGraphImpl.java
@@ -853,7 +853,7 @@ public class CHGraphImpl implements CHGraph, Storable<CHGraph> {
             if (weight > MAX_WEIGHT)
                 weightInt = MAX_WEIGHT_31;
             else
-                weightInt = Math.round((float) (weight * WEIGHT_FACTOR)) << 2;
+                weightInt = ((int) Math.round(weight * WEIGHT_FACTOR)) << 2;
             return weightInt;
         }
